+++ src/gui/animation/propertyvaluesholder.h
@@ -23,7 +23,7 @@ inline constexpr float lerp(float fromValue, float toValue, float fraction) {
     return float(fromValue * (1.f - fraction) + toValue * fraction);
 }
 inline int lerp(int startValue, int endValue, float fraction) {
-    return startValue + std::round(fraction * (endValue - startValue));
+    return int(startValue + std::round(fraction * (endValue - startValue)));
 }
 
 class Property{
