+++ src/gui/core/context.h
@@ -34,8 +34,8 @@ public:
     }
     virtual int getColor(const std::string&resid)=0;
     virtual int getDimension(const std::string&resid)=0;
-    virtual int getArray(const std::string&resname,std::vector<std::string>&)=0;
-    virtual int getArray(const std::string&resname,std::vector<int>&)=0;
+    virtual size_t getArray(const std::string&resname,std::vector<std::string>&)=0;
+    virtual size_t getArray(const std::string&resname,std::vector<int>&)=0;
     virtual ColorStateList* getColorStateList(const std::string&resid)=0;
     virtual AttributeSet obtainStyledAttributes(const std::string&resid)=0;
 };
