+++ src/gui/core/layout.cc
@@ -659,11 +659,11 @@ void  Layout::drawText(Canvas&canvas,int firstLine,int lastLine){
         canvas.show_text(u8line);
         if( (mCaretPos>=lineStart) && (mCaretPos<lineEnd) ){
             measureSize(line.substr(0,mCaretPos-lineStart),te,nullptr);
-            mCaretRect.left= x + te.x_advance;
-            mCaretRect.top = lineNum * mLineHeight;
+            mCaretRect.left= int(x + te.x_advance);
+            mCaretRect.top = int(lineNum * mLineHeight);
             mCaretRect.height= mLineHeight;
             measureSize(line.substr(mCaretPos-lineStart,1),te,nullptr);
-            mCaretRect.width = te.x_advance;
+            mCaretRect.width = int(te.x_advance);
         }
     }
 }
